+++ src/bk-user/bkuser/apis/open_web/pagination.py
@@ -14,7 +14,31 @@
 #
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
-from bkuser.common.pagination import NoCountCustomPageNumberPagination
+from rest_framework.response import Response
+
+from bkuser.common.pagination import CustomPageNumberPagination
+
+
+class NoCountCustomPageNumberPagination(CustomPageNumberPagination):
+    """
+    基于 CustomPageNumberPagination 实现的轻量分页器,用于不需要 count 的场景
+    """
+
+    def paginate_queryset(self, queryset, request, view=None):
+        self.request = request
+        page = self.get_page_number(request)
+        page_size = self.get_page_size(request)
+        if not page_size:
+            return None
+
+        offset = (page - 1) * page_size
+        return list(queryset[offset : offset + page_size])
+
+    def get_paginated_response(self, data):
+        return Response(data)
+
+    def get_paginated_response_schema(self, schema):
+        return schema
 
 
 def gen_no_count_pagination_class(max_page_size: int):
