+++ src/bk-user/bkuser/common/pagination.py
@@ -66,25 +66,3 @@ class CustomPageNumberPagination(PageNumberPagination):
                 "results": schema,
             },
         }
-
-
-class NoCountCustomPageNumberPagination(CustomPageNumberPagination):
-    """
-    基于 CustomPageNumberPagination 实现的轻量分页器,用于不需要 count 的场景
-    """
-
-    def paginate_queryset(self, queryset, request, view=None):
-        self.request = request
-        page = self.get_page_number(request)
-        page_size = self.get_page_size(request)
-        if not page_size:
-            return None
-
-        offset = (page - 1) * page_size
-        return list(queryset[offset : offset + page_size])
-
-    def get_paginated_response(self, data):
-        return Response(data)
-
-    def get_paginated_response_schema(self, schema):
-        return schema
